+++ actionview/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Rails 6.0.0.beta1 (January 18, 2019) ##
 
+*   ActionView::Template.finalize_compiled_template_methods is deprecated with
+    no replacement.
+
+*   config.action_view.finalize_compiled_template_methods is deprecated with
+    no replacement.
+
 *   Remove deprecated `image_alt` helper.
 
     *Rafael Mendonça França*